+++ src/main.js
@@ -23,7 +23,7 @@ new Vue({
   store,
   render: h => h(App),
   created(){
-    store.commit('addMenuToRouter',router) // 动态添加侧边栏菜单
+    store.dispatch('addMenuToRouter',router) // 动态添加侧边栏菜单
     store.commit('setting/changeTheme',  localStorage.getItem('theme') || 'dark') // 获取用户主题
     store.dispatch('setting/setLockScreen',  localStorage.getItem('lock-screen')) // 锁定屏幕
   }