+++ src/view/content/article-list.vue
@@ -160,6 +160,7 @@ import 'quill/dist/quill.core.css'
 import 'quill/dist/quill.snow.css'
 import 'quill/dist/quill.bubble.css'
 import dayjs from "dayjs";
+import { cleanParams } from "@/utils/helpers";
 
 const QUERY_PARAM = {
   page: 1,
@@ -226,10 +227,7 @@ export default {
   methods: {
     dayjs,
     async getList() {
-      const params = { ...this.queryParam }
-      Object.keys(params).forEach(k => {
-        if (params[k] === '' || params[k] === null || params[k] === undefined) delete params[k]
-      })
+      const params = cleanParams(this.queryParam)
       const res = await this.$api.getArticleList(params);
       this.tableData = res.list;
       this.total = res.total;