+++ src/view/notification/template-list.vue
@@ -137,6 +137,7 @@
 import FilterBar from "@/components/filter/FilterBar.vue";
 import FilterBarItem from "@/components/filter/FilterBarItem";
 import dayjs from "dayjs";
+import { cleanParams } from "@/utils/helpers";
 
 const QUERY_PARAM = {
   page: 1,
@@ -194,12 +195,7 @@ export default {
   },
   methods: {
     async getList() {
-      const params = { ...this.queryParam };
-      Object.keys(params).forEach((k) => {
-        if (params[k] === "" || params[k] === null || params[k] === undefined)
-          delete params[k];
-        if (Array.isArray(params[k]) && !params[k].length) delete params[k];
-      });
+      const params = cleanParams(this.queryParam);
       const res = await this.$api.getNotificationTemplateList(params);
       this.tableData = res.list;
       this.total = res.total;