+++ src/view/product/product-review.vue
@@ -180,6 +180,7 @@
 import FilterBar from "@/components/filter/FilterBar";
 import FilterBarItem from "@/components/filter/FilterBarItem";
 import dayjs from "dayjs";
+import { cleanParams } from "@/utils/helpers";
 
 const QUERY_PARAM = {
   page: 1,
@@ -217,10 +218,7 @@ export default {
   methods: {
     dayjs,
     async getList() {
-      const params = { ...this.queryParam }
-      Object.keys(params).forEach(k => {
-        if (params[k] === '' || params[k] === null || params[k] === undefined) delete params[k]
-      })
+      const params = cleanParams(this.queryParam)
       const res = await this.$api.getReviewList(params);
       this.tableData = res.list;
       this.total = res.total;